requirements.txt
anthropic>=0.40.0
watchdog>=4.0.0
pyyaml>=6.0
keyring>=25.0.0
